refactor: combine two handlers in one for login module
combine two handlers in one for login module

MODULE-1-PART-3-gallery
